Add inputs validations

+++ database/gasto.py
@@ -3,13 +3,57 @@ from datetime import datetime
 import sqlite3 # Type Hints
 
 # Agregar gasto
-def agregar_gasto(conn : sqlite3.Connection, telegram_usuario_id : int, categoria_id : int, comercio_id : int, monto : float, recibo_file_id : str, descripcion : str|None = None , fecha : str|None = None) -> int:
+def agregar_gasto(conn : sqlite3.Connection, telegram_usuario_id : int, categoria_id : int, comercio_id : int, monto : float,
+                recibo_file_id : str, descripcion : str|None = None , fecha : str|None = None) -> int:
     cursor = conn.cursor()
 
     # Validaciones 
-    if not fecha:
-        fecha = datetime.now().strftime("%Y-%m-%d %H:%M")
+    if type(telegram_usuario_id) is not int:
+        raise ValueError ("Coloque un telegram_usuario_id correcto.")
+    
+    if telegram_usuario_id <= 0:
+        raise ValueError ("Coloque un telegram_usuario_id correcto.")
 
+    if type(categoria_id) is not int:
+        raise ValueError ("Coloque una categoria_id correcta.")
+    
+    if categoria_id <= 0:
+        raise ValueError ("Coloque una categoria_id correcta.")
+
+    if type(comercio_id) is not int:
+        raise ValueError ("Coloque un comercio_id correcto.")
+    
+    if comercio_id <= 0:
+        raise ValueError ("Coloque un comercio_id correcto")
+    
+    if not isinstance(monto, (float)):
+        raise ValueError ("Coloque un monto correcto.")
+
+    if monto < 0.0:
+        raise ValueError ("Coloque un monto mayor a 0,0.")
+
+    if not isinstance(recibo_file_id, str):
+        raise ValueError ("Coloque el recibo_file_id correcto.")  
+
+    if not isinstance(descripcion, str):
+        raise ValueError ("Coloque una descripción correcta.") 
+
+    if not fecha:
+        fecha = datetime.now()
+
+    else:
+        try:
+            fecha = datetime.strptime(fecha, "%Y-%m-%d %H:%M")
+
+        except ValueError as V:
+            raise ValueError ("Coloque una fecha correcta. Ej: YYYY-MM-DD HH:MM") from V
+    
+    if 2025 > int(datetime.strftime(fecha, "%Y")):
+        raise ValueError ("Coloque un año correcto (2025+)")    
+
+    fecha = fecha.strftime("%Y-%m-%d %H:%M")
+
+    # SQL
     try:
         cursor.execute('''INSERT INTO GASTO 
                        (telegram_usuario_id, categoria_id, comercio_id, fecha, monto, descripcion, recibo_file_id)
@@ -27,8 +71,17 @@ def agregar_gasto(conn : sqlite3.Connection, telegram_usuario_id : int, categori
 def eliminar_gasto(conn : sqlite3.Connection, gasto_id : int):
     cursor = conn.cursor()
 
-    gasto = obtener_gasto(conn, gasto_id) # Validar registro
+    # Validaciones    
+    if type(gasto_id) is not int:
+        raise ValueError ("Coloque un gasto_id correcto.")
+    
+    if gasto_id <= 0:
+        raise ValueError ("Coloque un gasto_id correcto.")
+    
+    # Validar registro
+    obtener_gasto(conn, gasto_id) 
 
+    # SQL
     try:
         cursor.execute('''DELETE FROM GASTO
                        WHERE gasto_id = (?)''', (gasto_id, ))
@@ -40,13 +93,21 @@ def eliminar_gasto(conn : sqlite3.Connection, gasto_id : int):
     
 # Modificar gasto
 def modificar_gasto(conn : sqlite3.Connection, gasto_id : int, telegram_usuario_id : int|None = None, categoria_id : int|None = None, 
-                    comercio_id : int|None = None, fecha : str|None = None, monto : float|None = None, recibo_file_id : int|None = None, 
+                    comercio_id : int|None = None, fecha : str|None = None, monto : float|None = None, recibo_file_id : str|None = None, 
                     descripcion : str|None = None) -> int:
     cursor = conn.cursor()
 
-    gasto = obtener_gasto(conn, gasto_id) # Validar registro
+    # Validar gasto_id
+    if type(gasto_id) is not int:
+        raise ValueError ("Coloque un gasto_id correcto.")
 
-    # Validaciones 
+    if gasto_id <= 0:
+        raise ValueError ("Coloque un gasto_id correcto.")
+    
+    # Validar registro actual
+    gasto = obtener_gasto(conn, gasto_id) 
+
+    # Completar datos faltantes
     if telegram_usuario_id is None:
         telegram_usuario_id = gasto["telegram_usuario_id"]
 
@@ -62,12 +123,61 @@ def modificar_gasto(conn : sqlite3.Connection, gasto_id : int, telegram_usuario_
     if monto is None:
         monto = gasto["monto"]
 
-    if descripcion is None:
-        descripcion = gasto["descripcion"]
-        
     if recibo_file_id is None:
         recibo_file_id = gasto["recibo_file_id"]
 
+    if descripcion is None:
+        descripcion = gasto["descripcion"]
+    
+    # Validar datos finales
+    if type(telegram_usuario_id) is not int:
+        raise ValueError ("Coloque un telegram_usuario_id correcto.")
+    
+    if telegram_usuario_id <= 0:
+        raise ValueError ("Coloque un telegram_usuario_id correcto.")
+    
+
+    if type(categoria_id) is not int:
+        raise ValueError ("Coloque una categoria_id correcta.")
+    
+    if categoria_id <= 0:
+        raise ValueError ("Coloque una categoria_id correcta.")
+    
+
+    if type(comercio_id) is not int:
+        raise ValueError ("Coloque un comercio_id correcto.")
+    
+    if comercio_id <= 0:
+        raise ValueError ("Coloque un comercio_id correcto.")
+    
+
+    try:
+        fecha = datetime.strptime(fecha, "%Y-%m-%d %H:%M")
+
+    except ValueError as V:
+        raise ValueError ("Coloque una fecha correcta. Ej: YYYY-MM-DD HH:MM") from V
+    
+    if 2025 > int(datetime.strftime(fecha, "%Y")):
+        raise ValueError ("Coloque un año correcto (2025+)")    
+
+    fecha = fecha.strftime("%Y-%m-%d %H:%M")
+
+
+    if not isinstance(monto, float):
+        raise ValueError ("Coloque un monto correcto.")
+    
+    if monto <= 0.0:
+        raise ValueError ("Coloque un monto mayor a 0,0.")
+    
+
+    if not isinstance(descripcion, str):
+        raise ValueError ("Coloque una descripción correcta.")
+    
+
+    if type(recibo_file_id) is not str:
+        raise ValueError ("Coloque un recibo_file_id correcto.")
+    
+    # SQL
     try:
         cursor.execute('''UPDATE GASTO 
                        SET telegram_usuario_id = (?), categoria_id = (?), comercio_id = (?), fecha = (?),
@@ -86,6 +196,7 @@ def modificar_gasto(conn : sqlite3.Connection, gasto_id : int, telegram_usuario_
 def calcular_gasto_total(conn : sqlite3.Connection) -> float:
     cursor = conn.cursor()
     
+    # SQL
     try: 
         cursor.execute('''SELECT COALESCE(SUM(monto), 0.0) FROM GASTO
                        ''')
@@ -93,7 +204,7 @@ def calcular_gasto_total(conn : sqlite3.Connection) -> float:
         monto = cursor.fetchone()
 
     except sqlite3.Error as E:
-        raise Exception ("Error al sumar montos.") from E
+        raise Exception ("Error al sumar gastos.") from E
     
     return monto[0]
 
@@ -102,6 +213,9 @@ def calcular_gasto_mes(conn : sqlite3.Connection, mes : int, año : int|None = N
     cursor = conn.cursor()
     
     # Validaciones 
+    if not 0 < mes < 13: # Rango de meses inválidos
+        raise ValueError ("Coloque un número de mes correcto. Ej: (1-12)")
+
     if not año: # Año actual por defecto
         año_actual : str = datetime.now().strftime("%Y")
         año = año_actual
@@ -109,15 +223,13 @@ def calcular_gasto_mes(conn : sqlite3.Connection, mes : int, año : int|None = N
     if not 2025 <= año <= 2027: # Rango de años inválidos 
         raise ValueError("Año fuera del rango.")
 
-    if not 0 < mes < 13: # Rango de meses inválidos
-        raise ValueError ("Coloque un número de mes correcto. Ej: (1-12)")
-
     try:
         periodo = datetime.strptime(f"{año}-{mes:02}", "%Y-%m") 
 
     except ValueError as V:
         raise Exception ("Coloque un año correcto. Ej: AAAA") from V
 
+    # SQL
     try:
         cursor.execute('''
                        SELECT ROUND(COALESCE(SUM(monto), 0), 2) FROM GASTO
@@ -142,6 +254,7 @@ def calcular_gasto_comercio(conn : sqlite3.Connection, comercio_id : int) -> flo
     if comercio_id <= 0:
         raise ValueError ("Coloque un comercio_id correcto.")
 
+    # SQL
     try:
         cursor.execute('''
                        SELECT ROUND(COALESCE(SUM(monto), 0), 2) FROM GASTO WHERE comercio_id = (?)
@@ -165,6 +278,7 @@ def calcular_gasto_categoria(conn : sqlite3.Connection, categoria_id : int) -> f
     if categoria_id <= 0:
         raise ValueError("Coloque una categoria_id correcta.")
     
+    # SQL
     try:
         cursor.execute('''
                        SELECT ROUND(COALESCE(SUM(monto), 0), 2) FROM GASTO WHERE categoria_id = (?)
@@ -181,6 +295,7 @@ def calcular_gasto_categoria(conn : sqlite3.Connection, categoria_id : int) -> f
 def calcular_promedio_gastos(conn : sqlite3.Connection) -> float:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''
                     SELECT COALESCE(ROUND(AVG(monto), 2), 0.0) AS promedio FROM GASTO
@@ -197,6 +312,7 @@ def calcular_promedio_gastos(conn : sqlite3.Connection) -> float:
 def calcular_gasto_minimo(conn : sqlite3.Connection) -> float:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''SELECT COALESCE(ROUND(MIN(monto), 2), 0.0) AS monto from GASTO''')
 
@@ -211,6 +327,7 @@ def calcular_gasto_minimo(conn : sqlite3.Connection) -> float:
 def calcular_gasto_maximo(conn : sqlite3.Connection) -> float:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''SELECT COALESCE(ROUND(MAX(monto), 2), 0.0) AS monto from GASTO''')
 
@@ -225,6 +342,7 @@ def calcular_gasto_maximo(conn : sqlite3.Connection) -> float:
 def obtener_gasto(conn : sqlite3.Connection, gasto_id : int) -> dict:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''SELECT * FROM GASTO 
                        WHERE gasto_id = (?)
@@ -244,6 +362,7 @@ def obtener_gasto(conn : sqlite3.Connection, gasto_id : int) -> dict:
 def obtener_gastos(conn) -> list[dict]:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''SELECT * FROM GASTO''')
 
@@ -256,6 +375,7 @@ def obtener_gastos(conn) -> list[dict]:
 def obtener_totales_por_categoria(conn : sqlite3.Connection) -> list[dict]:
     cursor = conn.cursor()
 
+    # SQL
     try:
         cursor.execute('''
                        SELECT C.nombre, 
@@ -277,6 +397,7 @@ def obtener_totales_por_categoria(conn : sqlite3.Connection) -> list[dict]:
 def obtener_gastos_entre_fechas(conn : sqlite3.Connection, fecha_inicio : str, fecha_final : str) -> list[dict]:
     cursor = conn.cursor()
 
+    # Validaciones
     try:
         fecha_inicio = datetime.strptime(fecha_inicio, "%Y-%m-%d")
         fecha_final = datetime.strptime(fecha_final, "%Y-%m-%d")
@@ -284,7 +405,6 @@ def obtener_gastos_entre_fechas(conn : sqlite3.Connection, fecha_inicio : str, f
     except ValueError as V:
         raise Exception ("Coloque una fecha correcta.") from V
 
-    # Validaciones
     if not 2025 <= int(fecha_inicio.strftime("%Y")): # Años invalidos fecha_inicio
         raise ValueError ("Coloque un año correcto. Ej: YYYY.")
     
@@ -294,6 +414,7 @@ def obtener_gastos_entre_fechas(conn : sqlite3.Connection, fecha_inicio : str, f
     if fecha_inicio > fecha_final: # Diferencia entre fechas 
         raise ValueError ("Las fechas deben ser ascendentes.")
 
+    # SQL
     try:
         cursor.execute('''
                     SELECT * FROM GASTO WHERE DATE(fecha) BETWEEN (?) AND (?) ORDER BY fecha
@@ -314,9 +435,10 @@ def obtener_ultimos_gastos(conn : sqlite3.Connection, limit : int = 3) -> list[d
     if type(limit) is not int:
         raise ValueError ("Coloque un número.")
     
-    if 0 > limit:
-        raise ValueError ("Limit no puede ser negativo.")
+    if 2 > limit:
+        raise ValueError ("Limit debe ser mayor a 2.")
 
+    # SQL
     try:
         cursor.execute('''SELECT * FROM GASTO ORDER BY gasto_id DESC LIMIT (?)
                        ''', (limit, ))
@@ -337,6 +459,7 @@ def buscar_gastos_por_texto(conn : sqlite3.Connection, texto : str ="") -> list[
     if not texto.strip():
         raise ValueError ("Coloque un texto.")
 
+    # SQL
     try:
         cursor.execute('''
                        SELECT * FROM GASTO WHERE descripcion LIKE (?)
